requirements.txt
# Core dependencies
py-clob-client>=0.1.0
requests>=2.28.0
python-dotenv>=1.0.0

# HTTP retry logic
urllib3>=2.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
